refactor(edges): use EdgeWrapper instead of wrapEdge
